// agentx-frontend-plus/types/rag-dataset.ts
export interface CreateDatasetRequest {
  name: string;
  icon?: string;
  description?: string;
  embeddingModelId?: string;
}